// types.ts
export type PageType = 'privacy' | 'terms' | 'cookies' | 'about' | 'contact' | 'blog' | null;

export enum ToolCategory {
  ALL = 'All Tools',
  PDF = 'PDF & File',
  IMAGE = 'Image Tools',
  TEXT = 'Text & Encoding', // New Category
  SEO = 'SEO Tools',
  FINANCIAL = 'Financial',
  DATE_TIME = 'Date & Time',
  FITNESS = 'Fitness & Health',
  MATH_SCIENCE = 'Math & Science',
  UNIT_CONVERTER = 'Unit Converter',
  OTHER = 'Other Tools',
}